requirements.txt
asgiref==3.2.10
autopep8==1.5.4
bokeh==1.4.0
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
dj-database-url==0.5.0
Django==3.0
django-countries==6.1.3
django-crispy-forms==1.9.2
django-filter==2.3.0
django-i18n==1.0.4
django-js-asset==1.2.2
django-on-heroku==1.1.2
django-phonenumber-filter==1.0.0
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.4
packaging==20.4
phonenumbers==8.12.9
Pillow==7.2.0
po-excel-translate==3.0.1
polib==1.1.0
poxls==1.6.2
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycosat==0.6.3
PyMsgBox==1.0.9
pyparsing==2.4.7
pypi==2.1
python-dateutil==2.8.1
python-decouple==3.3
Python-dev==2.0.0.dev0
python-http-client==3.3.1
pytz==2020.1
PyYAML==5.3.1
requests==2.25.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
sendgrid==6.4.7
six==1.15.0
sqlparse==0.3.1
starkbank-ecdsa==1.1.0
toml==0.10.2
tornado==6.1
urllib3==1.26.2
Werkzeug==1.0.1
whitenoise==5.2.0
